common: Implement window filter for current display
As WNCK lacks the notion of Monitor as Implemented in Gdk,
this code fetches the geometry of WnckWindow, Fetches GdkMonitor
for that geometry, and checks against the plugins GdkMonitor.

In my opinion, it's a hack, which might be fixed with
libxfce4windowing, but for now, this is quite ok.

This keeps the new functionality disabled until it is implemented in
the respective plugins.

Signed-off-by: Mubashshir <ahmubashshir@gmail.com>